Add the support for command buffers

+++ ratchet-vulkan/ratchet-vulkan/VkDeviceMemory.cs
@@ -10,6 +10,6 @@ namespace Ratchet.Drawing.Vulkan
     [StructLayout(LayoutKind.Sequential)]
     public struct VkDeviceMemory
     {
-        internal IntPtr Handle;
+        internal UInt64 Handle;
     }
 }
